Change mana to be a float operator

+++ Assets/Mana/BlackMana.cs
@@ -1,14 +1,14 @@
 ﻿public class BlackMana
 {
-    public static void BurnBlack(int amount)
+    public static void BurnBlack(float amount)
     {
     }
 
-    public static void GainBlack(int amount)
+    public static void GainBlack(float amount)
     {
     }
 
-    public static Mana GetBase(int startingTotal = 0)
+    public static Mana GetBase(float startingTotal = 0)
     {
         return new Mana(ManaColor.Black, GainBlack, BurnBlack)
         {
