+++ Assets/Mana/RedMana.cs
@@ -1,14 +1,14 @@
 ﻿public class RedMana
 {
-    public static void BurnRed(int amount)
+    public static void BurnRed(float amount)
     {
     }
 
-    public static void GainRed(int amount)
+    public static void GainRed(float amount)
     {
     }
 
-    public static Mana GetBase(int startingTotal = 0)
+    public static Mana GetBase(float startingTotal = 0)
     {
         return new Mana(ManaColor.Red, GainRed, BurnRed)
         {
